% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{Frequencies}
\alias{Frequencies}
\title{Count annotation concept frequency}
\usage{
Frequencies(model, cui.list, cui.counts = NULL, Counts = NULL)
}
\arguments{
\item{model}{A model object like returned by \code{\link{loadDefaultModel}}.}

\item{cui.counts}{Counter of all drug concepts if Counts presented}

\item{Counts}{A frequency table to be added to. Default is Null.}

\item{cui1.list}{A list of vectors containing a sequence of drug concepts}
}
\value{
A data frame of the annotation concepts and their frequency (Counts).
}
\description{
Compute the annotation concept (of one kind) frequency of a model from a given list of drug concept vectors
}
