% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getDataSubmodel}
\alias{getDataSubmodel}
\title{Extract a subgraph of the model.}
\usage{
getDataSubmodel(model, verts)
}
\arguments{
\item{model}{A model object like returned by \code{\link{loadDefaultModel}}.}

\item{verts}{A vector containing a sequence of the concept identifiers that should be included in the returned subgraph.}
}
\value{
A new model containing only the requested concepts and their relationships.
}
\description{
Helper function for getting a submodel of the full data for creating
focused unit tests
}
