% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{Jaccard.F}
\alias{Jaccard.F}
\title{jaccard similarity based on frequency
TODO: Function description}
\usage{
Jaccard.F(F1, F2)
}
\description{
jaccard similarity based on frequency
TODO: Function description
}
