% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs.R
\name{listDrugAnnotations}
\alias{listDrugAnnotations}
\title{List related annotations}
\usage{
listDrugAnnotations(model, code, kinds = NULL)
}
\arguments{
\item{model}{A model object like returned by \code{\link{loadDefaultModel}}.}

\item{code}{A concept code (character vector) to look for.}

\item{kinds}{A vector of concept kinds to include in the returned list.}
}
\value{
A data frame of the matching annotation concepts.
}
\description{
Find all related concepts of a given concept, optionally restricted to set of concept kinds.
}
