% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs.R
\name{listRelatedDrugs}
\alias{listRelatedDrugs}
\title{List related drugs}
\usage{
listRelatedDrugs(model, code)
}
\arguments{
\item{model}{A model object like returned by \code{\link{loadDefaultModel}}.}

\item{code}{A concept code (character vector) to look for.}
}
\value{
A data frame of the matching drugs.
}
\description{
Identify all related drugs of a given concept of an ontology. This function is
equivalent to invoking \code{listRelatedNodes(model, code direction='down', kinds=DRUG_KIND)}.
}
