% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs.R
\name{listRelatedNodes}
\alias{listRelatedNodes}
\title{List related nodes}
\usage{
listRelatedNodes(model, code, direction = "down", kinds = "All")
}
\arguments{
\item{model}{A model object like returned by \code{\link{loadDefaultModel}}.}

\item{code}{A concept code (character vector) to look for.}

\item{direction}{A character vector containing one of \code{'up'}, \code{'down'},
or \code{'both'}.}

\item{kinds}{A vector of concept kinds to include in the returned list.}
}
\value{
A data frame of the matching nodes
}
\description{
Identify related nodes of a given concept of an ontology.
}
