% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{loadDefaultModel}
\alias{loadDefaultModel}
\title{Load the default model}
\usage{
loadDefaultModel()
}
\value{
A list with two elements: \code{graph} and \code{DefConcept}. The \code{DefConcept} element is a data frame
  of concepts; the \code{graph} element is an \code{igraph} of the relationships between concepts.
}
\description{
Load the default model
}
