% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{loadMeshTree}
\alias{loadMeshTree}
\title{Load a MeSH tree as a graph.}
\usage{
loadMeshTree(filename)
}
\arguments{
\item{filename}{The name of the file from which MeSH data should be loaded.}
}
\value{
TODO
}
\description{
TODO: Function description
}
