% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{loadRxNormMapping}
\alias{loadRxNormMapping}
\title{Load the RxNorm CUI <-> Concept table}
\usage{
loadRxNormMapping()
}
\value{
A data frame listing the correspondence between RxNorm CUIs and NDF-RT concepts.
}
\description{
Load the RxNorm CUI <-> Concept table
}
