% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{Similarity}
\alias{Similarity}
\title{Compute Similarity based on nodes}
\usage{
Similarity(model, cui1, cui2, weight = NULL, PD = F, View = F)
}
\arguments{
\item{model}{A model object like returned by \code{\link{loadDefaultModel}}.}

\item{cui1, cui2}{Vectors containing a sequence of drug concepts}

\item{weight}{A data.frame of of all notes' weights}

\item{PD}{A logical scalar. If PD is true, weight take power of 2.}

\item{View}{A logical scalar. If View is true, the drug lists will be visualized.}
}
\value{
a network similarity.
}
\description{
Compute node similarity between two drug lists. vectors containing a sequence of drug concepts
}
