% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viewModel}
\alias{viewModel}
\title{Visualize the model graph.}
\usage{
viewModel(model)
}
\arguments{
\item{model}{A model object like returned by \code{\link{loadDefaultModel}}.}
}
\description{
A naive model network visualization.
}
\details{
This function is a simple wrapper around the view.network function that accepts a model instead of a graph
}
